Added Items top interface
All Items now inherit from the master Item interface, if we need to
add more items, can easily add a new item to the items enum list,
them create a class that inherits from the interface. The rest of the
code was cleaned up to follow this. IMO, there's tones of improvements
that can be done, but I thought this would be a good start to making
this console app more "modular".